feat(#5): update navbar and add hamburger menu
