+++ package.json
@@ -9,15 +9,18 @@
   "dependencies": {
     "antd": "^5.7.3",
     "axios": "^1.4.0",
+    "body-scroll-lock": "^4.0.0-beta.0",
     "classname": "^0.0.0",
     "classnames": "^2.3.2",
     "eslint-config-next": "^13.4.10",
     "framer-motion": "^7.10.3",
     "google-map-react": "^2.2.1",
+    "hamburger-react": "^2.5.0",
     "next": "^13.0.5",
     "pure-react-carousel": "^1.30.1",
     "react": "^18.2.0",
     "react-awesome-slider": "^4.1.0",
+    "react-burger-menu": "^3.0.9",
     "react-dom": "^18.2.0",
     "react-google-forms-hooks": "^1.2.4",
     "react-icons": "^4.10.1",
